Add a new segment descriptor to GDT
This descriptor will be used to write in the video memory.

It will point to a data segment that begins at the linear
address 0xb8000. This is exactly where is located the video memory
when the text mode is set.

In fact, in such mode the video memory is 32Kb length. Despite we need
only 4Kb from it. This is because our screen is 25*80 characters. And
every character, to be displayed, needs 2 bytes. one to store its ASCII
code and the other to hold its attribute (its color).

that is we need exactly 4000 bytes in the video memory to display
25*80 = 2000 characters at once. So, 4Kb is more than enough!

So, the limit field of the new descriptor is equal to 1.
And its granularity bit is set to 1. that is: size=limit*G = 1*4Kb.

Remember that even addresses in the video memory are used to hold
the ASCII of a character to be displayed. Where odd addresses
are used to hold its attribute.


Signed-off-by: Issam Abdallah <iabdallah@yandex.com>