Move to user mode and execute a level-3 task
At this point, we begin diving in the details of the protected mode
features of the x86 processors. Therefore, our system become somewhat
complicated!

Instead of letting the kernel printing a character ('A') on the screen
in kernel mode (PL=0) we have delegated a user task to do that in user
mode (PL=3).

Before transferring the control to the task we have to move our
system from kernel mode to user mode!

One way to move to user mode is the execution of the IRET instruction.

Read section 'Exceptions and Interrupts' in the Intel 80386
manual to understand why we have clear the NT flag and why we have pushed
EIP, CS, EFLAGS, ESP and SS before executing IRET.

To use IRET we have to enable interrupts (using STI instruction)!

Before enabling interrupts we have to setup and initialize IDT. Otherwise,
we will end up with a crazy 'triple fault' upon the execution of STI!
The subprogram setup_idt (see linux-0.01/boot/head.s) will do that.

Implementing the multitasking feature of the INTEL x86 is very
complicated, at least for me, but I got it :)

A task must have its own TSS and LDT. TSS is used by the processor to
save the context of a task before switching to another and to retrieve
data and registers contents necessary to execute it efficiently. LDT
is used to separate the task's code and data from kernel and other tasks.

Read section 'Multitasking' in the manual to know how to create TSS, LDT
and transfer the control to a task (TR and LDTR).

The set_tssldt_des proc is taken from linux-0.01 source code with
some modifications. I used only the instructions to set the base fields
of the TSS and LDT descriptors.

Our task run in user mode, so we have to setup its own level 0 and 3 stacks,
that is, 'task_stack0' and 'task_stack'.

Unlike the last commit, the DPL field in the fourth descriptor of the GDT
is equal to 3! If we kept it equal to 0 (DPL=0), a protection exception
(and it will led to a 'triple fault') will be generated during
the execution of our task and exactly upon the loading of the 0x18 selector
in %gs. This is because our task will run in PL=3 and it is not permitted
by the protection mechanism to access data in level 0!


Signed-off-by: Issam Abdallah <iabdallah@yandex.com>