add a system call
What will happen with our last code snapshot (last commit),
if the DPL field of GDT's fourth descriptor become equal to 0 ?

It will not work and it will cause a triple fault! You can verify that.

In fact, any attempt to access the segment pointed by this descriptor from
level 3 (our task for example) will led to a protection exception,
immediately after loading a data segment  register by 0x18!
Because we no longer have the needed Privilege to access to it
from user mode!

To prevent this from happening, we have used system calls or Interrupts.

Interrupts and exceptions will create a safe hole in the processor's
protection mechanism, to enable us to call level 0 routines
(so is called a system call) and data from level 3 (user mode).

So, our task will not access directly to the VGA segment because it's no
longer allowed. Instead, it will use a system_call routine (IDT[0x80])
to do that safely.

Signed-off-by: Issam Abdallah <iabdallah@yandex.com>