Add timer interrupt and a scheduler
To add scheduling capability to our kernel, we have to program the
timer chip (the INTEL 8253 PIT) and add a timer interrupt handler.
So we did.

The timer is said to be the beating heart of the computer and the
scheduler is the soul of any OS.

At each timer interrupt, the kernel will switch from one task to another
using the LJMP instruction.

You can speed up or slow down the task switch by changing the HZ variable.

this is similar to linux-0.00, the first Linux version.

Enjoy :)

    This is it:)

Signed-off-by: Issam Abdallah <iabdallah@yandex.com>